Change `CrossBuildInMemoryCacheFactory` into an interface, add a test implementation and change several tests to use this.
