+++ subprojects/core/src/test/groovy/org/gradle/api/internal/changedetection/state/DefaultTaskHistoryStoreTest.groovy
@@ -20,9 +20,8 @@ import org.gradle.cache.CacheBuilder
 import org.gradle.cache.CacheRepository
 import org.gradle.cache.FileLockManager
 import org.gradle.cache.PersistentCache
-import org.gradle.cache.internal.CrossBuildInMemoryCacheFactory
+import org.gradle.cache.internal.TestCrossBuildInMemoryCacheFactory
 import org.gradle.cache.internal.filelock.LockOptionsBuilder
-import org.gradle.internal.event.DefaultListenerManager
 import spock.lang.Specification
 
 class DefaultTaskHistoryStoreTest extends Specification {
@@ -34,7 +33,7 @@ class DefaultTaskHistoryStoreTest extends Specification {
         PersistentCache backingCache = Mock()
 
         when:
-        new DefaultTaskHistoryStore(gradle, cacheRepository, new InMemoryCacheDecoratorFactory(false, new CrossBuildInMemoryCacheFactory(new DefaultListenerManager())))
+        new DefaultTaskHistoryStore(gradle, cacheRepository, new InMemoryCacheDecoratorFactory(false, new TestCrossBuildInMemoryCacheFactory()))
 
         then:
         1 * cacheRepository.cache(gradle, "taskHistory") >> cacheBuilder
